+++ react/IconBars.js
@@ -1,0 +1,3 @@
+import Bars from './components/icon/Bars/index'
+
+export default Bars